+++ src/pages/es/llms-full.txt.ts
@@ -1,0 +1,3 @@
+import { llmsEndpoint } from "../../endpoint-data";
+export const prerender = true;
+export const GET = () => llmsEndpoint("es", true);