+++ src/pages/es/resume.json.ts
@@ -1,0 +1,3 @@
+import { resumeEndpoint } from "../../endpoint-data";
+export const prerender = true;
+export const GET = () => resumeEndpoint("es");