+++ src/pages/es/search-index.json.ts
@@ -1,0 +1,3 @@
+import { searchEndpoint } from "../../endpoint-data";
+export const prerender = true;
+export const GET = () => searchEndpoint("es");